pyproject.toml
[project]
name = "arcane-mage"
version = "2.0.0"
description = "Automation helper for Fluxnodes"
readme = "README.md"
authors = [
    { name = "David White", email = "david@runonflux.io" }
]
requires-python = ">=3.13.0"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "cryptography>=46.0.5",
    "email-validator>=2.3.0",
    "keyring>=25.6.0,<25.7",
    "pydantic>=2.12.5",
    "pyescrypt>=0.1.0",
    "pyrage>=1.3.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
cli = [
    "click>=8.3.1",
]
tui = [
    "arcane-mage[cli]",
    "textual>=8.0.2",
]

[project.scripts]
arcane-mage = "arcane_mage.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arcane_mage"]

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
    "textual-dev>=1.8.0",
    "ty>=0.0.27",
    "types-aiofiles>=25.1.0.20251011",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ty.environment]
python-version = "3.13"
